pyproject.toml
[tool.poetry]
name = "python-backend-worker-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.6.1,<4"
celery = "5.0.5"
pycurl = "7.43.0.5"
boto3 = "1.16.52"
gevent = "20.12.1"
flask-restful = "0.3.8"
pyyaml = "5.4.1"
prometheus_client = "0.8.0"
sentry-sdk = "0.10.2"
requests = "2.25.1"
botocore = "1.19.52"
be-queue = {url = "https://pypi.infra.wish.com/api/package/be-queue/be-queue-0.1.0.tar.gz"}
hvac = "0.10.8"
opentracing-instrumentation = "3.3.1"
jaeger-client = "4.4.0"
fluent-logger = "0.10.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
mypy = "^0.800"
pylint = "^2.6.0"
pre-commit = "^2.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
